+++ index.js
@@ -1,4 +1,4 @@
+import './global.js'
 import Root from './Root.js'
 
-ReactDOM.render(
-  Root({ path: window.location.pathname }), document.getElementById('root'))
+ReactDOM.render(Root(), document.getElementById('root'))
